Fix failovertests to read params passed
keys_count, load_ratio and replica can now be passed to the
failovertests.
Eg.
test_failover_firewall,load_ratio=0.1

Reviewed-on: http://review.couchbase.org/9864
Tested-by: Karan Kumar <karan@couchbase.com>

+++ pytests/rebalancetests.py
@@ -215,7 +215,7 @@ class RebalanceBaseTest(unittest.TestCase):
             _keys_count = int(input.test_params['keys_count'])
         except KeyError:
             try:
-                _load_ratio = int(input.test_params['load_ratio'])
+                _load_ratio = float(input.test_params['load_ratio'])
             except KeyError:
                 _load_ratio = DEFAULT_LOAD_RATIO
